+++ src/cargo.rs
@@ -1,6 +1,6 @@
 use crate::config::*;
 use crate::errors::RunError;
-use crate::path_utils::get_source_walker;
+use crate::path_utils::{fix_unc_path, get_source_walker};
 use cargo_metadata::{diagnostic::DiagnosticLevel, CargoOpt, Message, Metadata, MetadataCommand};
 use lazy_static::lazy_static;
 use regex::Regex;
@@ -224,7 +224,8 @@ pub fn get_tests(config: &Config) -> Result<CargoOutput, RunError> {
             }
         }
     }
-    let manifest = config.manifest.as_path().to_str().unwrap_or("Cargo.toml");
+    let man_binding = config.manifest();
+    let manifest = man_binding.as_path().to_str().unwrap_or("Cargo.toml");
     let metadata = MetadataCommand::new()
         .manifest_path(manifest)
         .features(CargoOpt::AllFeatures)
@@ -280,7 +281,7 @@ fn run_cargo(
                         }
                         result
                             .test_binaries
-                            .push(TestBinary::new(PathBuf::from(path), ty));
+                            .push(TestBinary::new(fix_unc_path(path.as_std_path()), ty));
                         package_ids.push(Some(art.package_id.clone()));
                     }
                 }
@@ -343,7 +344,7 @@ fn run_cargo(
                 res.cargo_dir = package
                     .manifest_path
                     .parent()
-                    .map(|x| PathBuf::from(x.to_path_buf()));
+                    .map(|x| fix_unc_path(x.as_std_path()));
                 res.pkg_name = Some(package.name.clone());
                 res.pkg_version = Some(package.version.to_string());
                 res.pkg_authors = Some(package.authors.clone());
@@ -375,7 +376,7 @@ fn run_cargo(
         let should_panics = get_attribute_candidates(&dir_entries, config, "should_panic");
         let no_runs = get_attribute_candidates(&dir_entries, config, "no_run");
         for dt in &dir_entries {
-            let mut tb = TestBinary::new(dt.path().to_path_buf(), ty);
+            let mut tb = TestBinary::new(fix_unc_path(dt.path()), ty);
 
             if let Some(meta) = DocTestBinaryMeta::new(dt.path()) {
                 if no_runs
@@ -393,7 +394,7 @@ fn run_cargo(
             let mut current_dir = dt.path();
             loop {
                 if current_dir.is_dir() && current_dir.join("Cargo.toml").exists() {
-                    tb.cargo_dir = Some(current_dir.to_path_buf());
+                    tb.cargo_dir = Some(fix_unc_path(current_dir));
                     break;
                 }
                 match current_dir.parent() {
@@ -671,27 +672,10 @@ fn handle_llvm_flags(value: &mut String, config: &Config) {
     }
 }
 
-pub fn rustdoc_flags(config: &Config) -> String {
-    const RUSTDOC: &str = "RUSTDOCFLAGS";
-    let common_opts = " -Cdebuginfo=2 --cfg=tarpaulin ";
-    let mut value = format!(
-        "{} --persist-doctests {} -Zunstable-options ",
-        common_opts,
-        config.doctest_dir().display()
-    );
-    if let Ok(vtemp) = env::var(RUSTDOC) {
-        if !vtemp.contains("--persist-doctests") {
-            value.push_str(vtemp.as_ref());
-        }
-    }
-    handle_llvm_flags(&mut value, config);
-    deduplicate_flags(&value)
-}
-
-fn look_for_rustflags_in_table(value: &Value) -> String {
+fn look_for_field_in_table(value: &Value, field: &str) -> String {
     let table = value.as_table().unwrap();
 
-    if let Some(rustflags) = table.get("rustflags") {
+    if let Some(rustflags) = table.get(field) {
         if rustflags.is_array() {
             let vec_of_flags: Vec<String> = rustflags
                 .as_array()
@@ -712,42 +696,42 @@ fn look_for_rustflags_in_table(value: &Value) -> String {
     }
 }
 
-fn look_for_rustflags_in_file(path: &Path) -> Option<String> {
+fn look_for_field_in_file(path: &Path, section: &str, field: &str) -> Option<String> {
     if let Ok(contents) = read_to_string(path) {
         let value = contents.parse::<Value>().ok()?;
 
-        let rustflags_in_file: Vec<String> = value
+        let value: Vec<String> = value
             .as_table()?
             .into_iter()
             .map(|(s, v)| {
-                if s.as_str() == "build" {
-                    look_for_rustflags_in_table(v)
+                if s.as_str() == section {
+                    look_for_field_in_table(v, field)
                 } else {
                     String::new()
                 }
             })
             .collect();
 
-        Some(rustflags_in_file.join(" "))
+        Some(value.join(" "))
     } else {
         None
     }
 }
 
-fn look_for_rustflags_in(path: &Path) -> Option<String> {
+fn look_for_field_in_section(path: &Path, section: &str, field: &str) -> Option<String> {
     let mut config_path = path.join("config");
 
-    let rustflags = look_for_rustflags_in_file(&config_path);
-    if rustflags.is_some() {
-        return rustflags;
+    let value = look_for_field_in_file(&config_path, section, field);
+    if value.is_some() {
+        return value;
     }
 
     config_path.pop();
     config_path.push("config.toml");
 
-    let rustflags = look_for_rustflags_in_file(&config_path);
-    if rustflags.is_some() {
-        return rustflags;
+    let value = look_for_field_in_file(&config_path, section, field);
+    if value.is_some() {
+        return value;
     }
 
     None
@@ -761,14 +745,18 @@ fn build_config_path(base: impl AsRef<Path>) -> PathBuf {
     config_path
 }
 
-fn gather_config_rust_flags(config: &Config) -> String {
-    if let Some(rustflags) = look_for_rustflags_in(&build_config_path(&config.root())) {
-        return rustflags;
+fn gather_config_field_from_section(config: &Config, section: &str, field: &str) -> String {
+    if let Some(value) =
+        look_for_field_in_section(&build_config_path(&config.root()), section, field)
+    {
+        return value;
     }
 
     if let Ok(cargo_home_config) = env::var("CARGO_HOME") {
-        if let Some(rustflags) = look_for_rustflags_in(&PathBuf::from(cargo_home_config)) {
-            return rustflags;
+        if let Some(value) =
+            look_for_field_in_section(&PathBuf::from(cargo_home_config), section, field)
+        {
+            return value;
         }
     }
 
@@ -792,9 +780,29 @@ pub fn rust_flags(config: &Config) -> String {
     if let Ok(vtemp) = env::var(RUSTFLAGS) {
         value.push_str(&DEBUG_INFO.replace_all(&vtemp, " "));
     } else {
-        let vtemp = gather_config_rust_flags(config);
+        let vtemp = gather_config_field_from_section(config, "build", "rustflags");
         value.push_str(&DEBUG_INFO.replace_all(&vtemp, " "));
     }
+    deduplicate_flags(&value)
+}
+
+pub fn rustdoc_flags(config: &Config) -> String {
+    const RUSTDOC: &str = "RUSTDOCFLAGS";
+    let common_opts = " -Cdebuginfo=2 --cfg=tarpaulin ";
+    let mut value = format!(
+        "{} --persist-doctests {} -Zunstable-options ",
+        common_opts,
+        config.doctest_dir().display()
+    );
+    if let Ok(vtemp) = env::var(RUSTDOC) {
+        if !vtemp.contains("--persist-doctests") {
+            value.push_str(vtemp.as_ref());
+        }
+    } else {
+        let vtemp = gather_config_field_from_section(config, "build", "rustdocflags");
+        value.push_str(&vtemp);
+    }
+    handle_llvm_flags(&mut value, config);
     deduplicate_flags(&value)
 }
 
@@ -871,7 +879,7 @@ mod tests {
         };
 
         assert_eq!(
-            look_for_rustflags_in_table(&list_flags),
+            look_for_field_in_table(&list_flags, "rustflags"),
             "--cfg=foo --cfg=bar"
         );
 
@@ -880,7 +888,7 @@ mod tests {
         };
 
         assert_eq!(
-            look_for_rustflags_in_table(&string_flags),
+            look_for_field_in_table(&string_flags, "rustflags"),
             "--cfg=bar --cfg=baz"
         );
     }